% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-matrix.R
\name{as_mixing_matrix}
\alias{as_mixing_matrix}
\title{Construct a graph object's mixing matrix.}
\usage{
as_mixing_matrix(g, row_attr, col_attr = row_attr, direction = c("all",
  "out", "in"))
}
\arguments{
\item{g}{\code{<igraph>}, \code{<network>}}

\item{row_attr}{Either the name of a node/vertex attribute or vector of same
length as the number of nodes in \code{g}.}

\item{col_attr}{If not provided, the same as \code{row_attr}. Otherwise, a vector
of the same length as the number of nodes in \code{g}.}

\item{direction}{\code{<chr>}  The direction of ties to take into a account.
\itemize{
\item \code{"all"}, \code{"out"}, or \code{"in"}
}}
}
\description{
Using a categorical vertex attribute, construct a matrix depicting network mixing
patterns.
}
\examples{
as_mixing_matrix(jemmah_islamiyah, row_attr = "role")

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
