% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{assort_discrete}
\alias{assort_discrete}
\alias{assort_scalar}
\alias{assort_degree}
\title{Newmans' Assortativity Coefficient}
\usage{
assort_discrete(g, node_attr_name, ...)

assort_scalar(g, val1, val2 = NULL)

assort_degree(g)
}
\arguments{
\item{g}{\code{<igraph>}, \code{<network>}}

\item{node_attr_name}{\code{<chr>}  The name of a node/vertex attribute in \code{g}.}

\item{...}{Arguments passed to or from other methods.}

\item{val1}{Continuous vector.}

\item{val2}{Continuous vector.}
}
\description{
Given a categorical vertex attribute describing mutually exclusive groups,
the assortativity coefficient represents the tendency for nodes to form ties
with similar alters.
}
\details{
\deqn{ r = \frac{ \sum_i{e_{ii} - \sum_i{a_ib_i}} } {1 - \sum_i{a_ib_i} } }
\itemize{
\item \eqn{e}: mixing matrix
\item \eqn{e_{ii}}: diagonal of mixing matrix
\item \eqn{a_i}: row sums of mixing matrix
\item \eqn{b_i}: column sums of mixing matrix
}
}
\examples{
data("faux.desert.high", package = "ergm")
assort_discrete(faux.desert.high, node_attr_name = "race")


data("sampson", package = "ergm")
assort_discrete(samplike, "group")

}
\references{
M. E. J. Newman, Mixing patterns in networks, Physical Review E,
67 026126, 2003. \url{https://arxiv.org/pdf/cond-mat/0209450.pdf}.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
