% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_index.R
\name{ei_index}
\alias{ei_index}
\title{Krackhardt's E-I Index}
\usage{
ei_index(g, node_attr_name, scope = c("global", "group", "node"), ...)
}
\arguments{
\item{g}{\code{<igraph>}, \code{<network>}}

\item{node_attr_name}{\code{<chr>}  The name of a node/vertex attribute in \code{g}.}

\item{scope}{\code{<chr>} The target scope of the resulting EI-Index
\itemize{
\item \code{"global"}, \code{"group"}, or \code{"node"}
}}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Given a categorical vertex attribute describing mutually exclusive groups, the E-I
index represents a ratio of external to internal ties.
}
\details{
\deqn{E\mbox{-}I~Index = \frac{EL-IL}{EL+IL}}
\itemize{
\item \eqn{EL}: external ties
\item \eqn{IL}: internal ties
}
}
\examples{
# <igraph>
ei_index(jemmah_islamiyah, node_attr_name = "role")
ei_index(jemmah_islamiyah, node_attr_name = "role", scope = "group")
ei_index(jemmah_islamiyah, node_attr_name = "role", scope = "node")

# <network>
data("sampson", package = "ergm")
ei_index(samplike, node_attr_name = "group")
ei_index(samplike, node_attr_name = "group", scope = "group")
ei_index(samplike, node_attr_name = "group", scope = "node")

}
\references{
Krackhardt, David, and Robert N. Stern. "Informal Networks and
Organizational Crises: An Experimental Simulation." Social Psychology Quarterly 51, no.
2 (1988): 123-40. \url{http://www.jstor.org/stable/2786835}.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
