% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jemmah_islamiyah.R
\docType{data}
\name{jemmah_islamiyah}
\alias{jemmah_islamiyah}
\title{Koschade's Jemaah Islamiyah (2002 Bali Bombings)}
\format{An object of class \code{igraph} of length 10.}
\source{
\url{https://www.tandfonline.com/doi/abs/10.1080/10576100600798418}
}
\usage{
jemmah_islamiyah
}
\description{
\code{jemmah_islamiyah} is a reconstruction of the network analyzed
in Stuart Koschade's A Social Network Analysis of Jemaah Islamiyah.
}
\references{
Koschade, Stuart (2006) A Social Network Analysis of Jemaah
Islamiyah: The Applications to Counter-Terrorism and Intelligence. Studies in
Conflict and Terrorism Vol. 29(6):pp. 559-575
}
\keyword{datasets}
