% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_mixing_matrix}
\alias{as_mixing_matrix}
\alias{as_mixing_matrix.igraph}
\alias{as_mixing_matrix.network}
\title{Construct a graph object's mixing matrix.}
\usage{
as_mixing_matrix(g, node_attr_name)

\method{as_mixing_matrix}{igraph}(g, node_attr_name)

\method{as_mixing_matrix}{network}(g, node_attr_name)
}
\arguments{
\item{g}{\itemize{
\item \code{<igraph>}
\itemize{
\item \code{<network>}, if \code{{intergraph}} is installed.
}
}}

\item{node_attr_name}{\itemize{
\item \code{<chr>}
\item The name of a node/vertex attribute in \code{g}.
}}
}
\description{
Using a categorical vertex attribute, construct a matrix depicting network mixing
patterns.
}
\examples{
as_mixing_matrix(jemmah_islamiyah, node_attr_name = "role")

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
